+++ spec/base_spec_helper.rb
@@ -7,13 +7,6 @@ begin
 rescue LoadError
 end
 
-if ENV["COVER"]
-  require "simplecov"
-  SimpleCov.root File.join(File.dirname(__FILE__), "..")
-  SimpleCov.add_filter "/spec/"
-  SimpleCov.start
-end
-
 PROJECT_ROOT = File.expand_path("../", __dir__)
 
 RSpec.configure do |config|